Expose some elements of MKNetworkOperation so that they can be referenced from subclasses.
This adds connection:willSendRequest:redirectResponse: and
connectionDidFinishLoading: to its public interface, and moves
connection, request, and response from the private to the public.

This allows subclasses of MKNetworkOperation to override
those methods while still calling the base class.  It also allows
connection, request, and response to be referenced from those overrides.