Use mutableCopy for many of the fields in copyForRetry.
These fields are declared as NSMutableArray / NSMutableDictionary,
so it's violating the type safety to use plain copy.

This is important if you try to cancel an operation that came from
cancelForRetry.  [MKNetworkOperation cancel] will clear all the
lists of all the callbacks, which will crash if the lists are not
mutable.