Include the network status in the kReachabilityChangedNotification userInfo.
The current API requires the handler to call back in to this class to
get this info, which means making another SCNetworkReachabilityGetFlags
call.  This is pointless -- we have the flags at the point we send the
notification.

This also fixes a retain cycle through self in the dispatch block
for the postNotification call.