Add addJSONObject.
This allows the caller to specify an object that should be serialized as
JSON as the body of the request.

We already have MKNKPostDataEncodingTypeJSON, but this can only send a
JSON dictionary, because it uses the params given to initWithURLString, which
is declared as an NSDictionary.  addJSONObject lets you provide an array
or a primitive as the body.