Make the onNotModified handler be MKNKResponseBlock, not MKNKVoidBlock.
This means that the operation is passed to the callback, so if the
callback wants to inspect the headers on the 304 response, it can do so.

It's actually impossible for the handler to hold its own reference to the
operation, because that reference will be for the wrong object if
the operation is retried (which implies that it is copied).