Implement a feature whereby some network calls do not cause the application network activity indicator to spin.
This is useful for "long-poll" operations where the network call
is intended to block on the server side until something happens.
You don't want the activity indicator spinning for these, because
it will effectively never stop.

To do this, add MKNetworkOperation.hideActivityIndicator, which is
set by the caller.  Also add an activityThreshold field to
NSOperationQueue, through a category with an associated object.
This is incremented / decremented by QueueMonitorJob whenever a
hideActivityIndicator job is added to the queue.

As part of this change, we now create the QueueMonitorJob before
adding the NSOperation to the NSOperationQueue.  We need this to
be that way around because the act of adding to the queue causes
a KVO notification on operationCount, and that's when we're calling
checkActivity and updateNetworkActivityIndicator.