Fix the handling of the network activity spinner in the presence of long-poll operations and a batch of other network activity.
We were removing the latest entries from QueueMonitor.jobs when the
list got too long, to keep it trimmed.  This was done without regard
of whether or not the operation was complete.  That's fine if all
operations are taking more or less the same time, but for long-poll
operations, this meant that we would throw away the QueueMonitorJob
for that operation before it actually completed.  This meant that
we didn't correctly adjust the networkActivityThreshold for that
queue, and so we would show the activity indicator even when there were
only long-poll operations left in progress.

This fixed itself once the flurry of network activity went away, because
it only happened if the long-poll operation was still pending and it
became 50th or greater in terms of recent operations performed (i.e.
it got to the bottom of QueueMonitor.jobs).